Standard-suite speedup: cut ~47 model loads to ~30, add quick modes, parallel CPU phase
- PPL oracle batch mode (10 loads -> 1), golden-run QUICK (15 -> 8 loads),
  golden->llama output reuse, kvbatch --pairs single-process (5 -> 1 loads
  via Graph::reinit_kv), kvctx single shared Graph
- Env-gated quick trims (defaults unchanged): MATMUL/MTP/KVCTX/REGRESSION/
  KVQUANT/ROPE_QUICK, BATCH_QUICK; quant-precision row defaults 2048 -> 128;
  dequant nblocks 16 -> 4 with shared buffers; rope fused D2H
- check_all.sh phase A parallel with per-step logs; --quick propagates safe
  quick envs + server quick flag; rdna4_gates build target; ccache for CXX
- CPU gates: server-http 3.1s -> 2.3s, sampler 194ms -> 61ms, loader 17MB
  head read -> 16B; check_target.sh line-99 gpu-literal bug fixed
- Full rdna4_gates build clean; CPU half PASS

+++ include/rdna4/server.h
@@ -96,7 +96,8 @@ struct ServerOptions {
   std::string host = "127.0.0.1";
   int port = 8080;
   int backlog = 16;
-  int recv_timeout_s = 30;   // one request's header+body, per recv() call
+  double recv_timeout_s = 30.0;  // one request's header+body, per recv() call
+                                   // (fractional allowed: 0.2 = 200 ms, used by tests)
   int send_timeout_s = 120;  // per send() call
   std::size_t max_body = 8u << 20;
   bool verbose = false;
